VNC Launch & VM Window Update
Updated VM Node Window to have a menu to control start/stop/vnc of that VNC. It now luanchs the VNC console but issues connecting.
VM Window now shows the VM OS label and type. There's a icon for the OS, but need to update this based on the type. It's current set for Windows. Oh yeah, I'm using icons again.